Remove Core project and restructure Domain.
The Core project, including its configuration and code, was removed entirely.

Domain, where new classes (User, TodoStatus) were added, and TodoItem was reorganized.

Updated the solution structure to include a new Blazor project.
