--- Core/Core.http
@@ -1,6 +1,0 @@
-@Core_HostAddress = http://localhost:5225
-
-GET {{Core_HostAddress}}/weatherforecast/
-Accept: application/json
-
-###
