+++ Domain/Domain.csproj
@@ -4,6 +4,7 @@
         <TargetFramework>net9.0</TargetFramework>
         <Nullable>enable</Nullable>
         <ImplicitUsings>enable</ImplicitUsings>
+        <RootNamespace>Domain</RootNamespace>
     </PropertyGroup>
 
 </Project>
