+++ Domain/Todo/TodoStatus.cs
@@ -1,0 +1,10 @@
+namespace Domain.Todo;
+
+public enum TodoStatus
+{
+    None = 0,
+    Due = 1,
+    Overdue = 2,
+    Completed = 3,
+    Deleted = 99
+}
