+++ TodoApp.sln
@@ -4,6 +4,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Api", "Api\Api.csproj", "{4
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Domain", "Domain\Domain.csproj", "{5767A5A1-990C-4B5E-836D-46356C74BFA6}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Blazor", "Blazor\Blazor.csproj", "{E80B7DFA-FCEB-424A-83E7-AECDBF6CFD35}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -18,5 +20,9 @@ Global
 		{5767A5A1-990C-4B5E-836D-46356C74BFA6}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{5767A5A1-990C-4B5E-836D-46356C74BFA6}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{5767A5A1-990C-4B5E-836D-46356C74BFA6}.Release|Any CPU.Build.0 = Release|Any CPU
+		{E80B7DFA-FCEB-424A-83E7-AECDBF6CFD35}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{E80B7DFA-FCEB-424A-83E7-AECDBF6CFD35}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{E80B7DFA-FCEB-424A-83E7-AECDBF6CFD35}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{E80B7DFA-FCEB-424A-83E7-AECDBF6CFD35}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 EndGlobal
